added ability to download files from modac.cancer.gov

+++ common/file_utils.py
@@ -39,6 +39,50 @@ else:
     from six.moves.urllib.request import urlretrieve
 
 
+modac_user = None
+modac_pass = None
+def get_file_from_modac(fname, origin):
+    """ Downloads a file from the "Model and Data Clearning House" (MoDAC)
+    repository. Users should already have a MoDAC account to download the data.
+    Accounts can be created on modac.cancer.gov
+
+        Parameters
+        ----------
+        fname : string
+            path on disk to save the file
+        origin : string
+            original MoDAC URL of the file
+
+        Returns
+        ----------
+        Path to the downloaded file      
+    """
+
+    print('Downloading data from modac.cancer.gov, make sure you have an account first.')
+
+    global modac_user
+    global modac_pass
+    if modac_user is None: 
+        modac_user = input("MoDaC Username: ")
+
+    if modac_pass is None:
+        import getpass 
+        modac_pass = getpass.getpass("MoDaC Password: ")
+
+    data = {}
+    headers = {'Content-Type': 'application/json'}
+    auth = (modac_user, modac_pass)
+    url = origin + '/download'
+    print(url)
+    r = requests.post(url, data = data, headers = headers, auth = auth)
+    r.raise_for_status()
+    print(r.status_code)
+    print('Saving file to:' + fname)
+    out_file = open(fname, "wb")
+    out_file.write(r.content)
+    out_file.close()
+    return fname
+
 def get_file(fname, origin, untar=False,
              # md5_hash=None, datadir='../Data/common'):
              # md5_hash=None, cache_subdir='common', datadir='../Data/common'):
@@ -112,32 +156,35 @@ def get_file(fname, origin, untar=False,
     '''
 
     if download:
-        print('Downloading data from', origin)
-        global progbar
-        progbar = None
-
-        def dl_progress(count, block_size, total_size):
+        if 'modac.cancer.gov' in origin:
+            get_file_from_modac(fpath, origin)
+        else:
+            print('Downloading data from', origin)
             global progbar
-            if progbar is None:
-                progbar = Progbar(total_size)
-            else:
-                progbar.update(count * block_size)
-
-        error_msg = 'URL fetch failure on {}: {} -- {}'
-        try:
+            progbar = None
+    
+            def dl_progress(count, block_size, total_size):
+                global progbar
+                if progbar is None:
+                    progbar = Progbar(total_size)
+                else:
+                    progbar.update(count * block_size)
+    
+            error_msg = 'URL fetch failure on {}: {} -- {}'
             try:
-                urlretrieve(origin, fpath, dl_progress)
-                # fpath = wget.download(origin)
-            except URLError as e:
-                raise Exception(error_msg.format(origin, e.errno, e.reason))
-            except HTTPError as e:
-                raise Exception(error_msg.format(origin, e.code, e.msg))
-        except (Exception, KeyboardInterrupt) as e:
-            if os.path.exists(fpath):
-                os.remove(fpath)
-            raise
-        progbar = None
-        print()
+                try:
+                    urlretrieve(origin, fpath, dl_progress)
+                    # fpath = wget.download(origin)
+                except URLError as e:
+                    raise Exception(error_msg.format(origin, e.errno, e.reason))
+                except HTTPError as e:
+                    raise Exception(error_msg.format(origin, e.code, e.msg))
+            except (Exception, KeyboardInterrupt) as e:
+                if os.path.exists(fpath):
+                    os.remove(fpath)
+                raise
+            progbar = None
+            print()
 
     if untar:
         if not os.path.exists(untar_fpath):
